requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.0
PyNaCl>=1.5.0
aiohttp>=3.9.0